Add Makefile and Caddy config generator
This adds some tools and helpers to make deploying a new instance
easier. I also encourage the use of Caddy, which allows reverse
proxy of the required paths, and a simple webserver for anything
that doesn't match anything known about by the API instance.

There is a very simple index.html which is copied to the webroot
folder (/var/lib/relay/webroot) which the end user can then modify
if wanted.

This consists of squashed commits:
    Put the default webroot in a sensible location
    Add a slightly better 404 string
    Add caddyfile generator
    Add config hooks
    Add basic Makefile

See https://git.artemai.art/xrobau/Activity-Relay for raw commits

Signed-Off-By: Rob Thomas <xrobau@gmail.com>